feat: move WebSocket to independent package
